mem: more fix and updates

+++ src/mem/mem_interface.hh
@@ -543,7 +543,7 @@ class DRAMInterface : public MemInterface
          * Incremented when a refresh event is started as well
          * Used to determine when a low-power state can be entered
          */
-        uint8_t outstandingEvents;
+        unsigned outstandingEvents;
 
         /**
          * delay low-power exit until this requirement is met
